Combined the code in ragas_.py with langfuse_ragas.py and deleted ragas_.py
